% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{count_kmers}
\alias{count_kmers}
\title{Count K-mers in a DNA Sequence}
\usage{
count_kmers(sequence, k, alphabet = c("A", "C", "G", "T"))
}
\arguments{
\item{sequence}{A character string, Biostrings XString, or DNAStringSet}

\item{k}{Integer. Length of k-mers to count}

\item{alphabet}{Character vector. Default: c("A", "C", "G", "T")}
}
\description{
Count K-mers in a DNA Sequence
}
