% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_motifs.R
\name{.build_class_dataframe}
\alias{.build_class_dataframe}
\title{Build Class Data.frame for Motif Analysis}
\usage{
.build_class_dataframe(motif_cluster, motif_names, cluster_cols)
}
\description{
Build Class Data.frame for Motif Analysis
}
\keyword{internal}
