% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_motifs.R
\name{.calculate_motif_scores}
\alias{.calculate_motif_scores}
\title{Calculate Motif Importance Scores Across Clusters}
\usage{
.calculate_motif_scores(classe_df, cluster_cols, motif_cluster)
}
\description{
Calculate Motif Importance Scores Across Clusters
}
\keyword{internal}
