% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_motifs.R
\name{.evaluate_motif_importance}
\alias{.evaluate_motif_importance}
\title{Evaluate Motif Importance Across Multiple Clusters}
\usage{
.evaluate_motif_importance(row, cluster_cols, motif_cluster)
}
\description{
Evaluate Motif Importance Across Multiple Clusters
}
\keyword{internal}
