% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models_rf_xgboost.R
\name{.is_positive_integer}
\alias{.is_positive_integer}
\title{Check if value is a positive integer}
\usage{
.is_positive_integer(x)
}
\description{
Check if value is a positive integer
}
\keyword{internal}
