% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sample_train_validation.R,
%   R/seq_classification.R
\name{.select_by_class}
\alias{.select_by_class}
\title{Select k sequences per class}
\usage{
.select_by_class(data, k, dataset_name = "")

.select_by_class(data, k, dataset_name = "")
}
\description{
Select k sequences per class

Select k sequences per class
}
\keyword{internal}
