% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models_rf_xgboost.R
\name{.validate_columns}
\alias{.validate_columns}
\title{Validate required columns}
\usage{
.validate_columns(data, required_cols, dataset_name)
}
\description{
Validate required columns
}
\keyword{internal}
