% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{generate_all_kmers}
\alias{generate_all_kmers}
\title{Generate all possible k-mers from an alphabet}
\usage{
generate_all_kmers(alphabet, k)
}
\arguments{
\item{alphabet}{Character vector of nucleotides}

\item{k}{Integer. Length of k-mers}
}
\value{
Character vector of all possible k-mers
}
\description{
Generate all possible k-mers from an alphabet
}
