% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{head.kmer_counts}
\alias{head.kmer_counts}
\title{Head method for kmer_counts objects}
\usage{
\method{head}{kmer_counts}(x, n = 6L, ...)
}
\arguments{
\item{x}{A kmer_counts object}

\item{n}{Number of elements to return}

\item{...}{Additional arguments (ignored)}
}
\description{
Head method for kmer_counts objects
}
