% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dendrogram.R
\name{print.dendrogram_result}
\alias{print.dendrogram_result}
\title{Print method for dendrogram_result objects}
\usage{
\method{print}{dendrogram_result}(x, ...)
}
\arguments{
\item{x}{A dendrogram_result object created by \code{\link{create_dendrogram}}.}

\item{...}{Additional arguments (not currently used).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Displays a summary of the dendrogram analysis result.
}
