% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{print.kmer_counts}
\alias{print.kmer_counts}
\title{Print method for kmer_counts objects}
\usage{
\method{print}{kmer_counts}(x, ...)
}
\arguments{
\item{x}{A kmer_counts object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for kmer_counts objects
}
