% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{print.kmer_data}
\alias{print.kmer_data}
\title{Print Method for kmer_data Objects}
\usage{
\method{print}{kmer_data}(x, ...)
}
\arguments{
\item{x}{A kmer_data object created by \code{\link{create_data}}.}

\item{...}{Additional arguments (not currently used).}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Displays a summary of the k-mer data created by \code{\link{create_data}}.
}
