% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_motifs.R
\name{print.select_motifs}
\alias{print.select_motifs}
\title{Print Method for select_motifs Objects}
\usage{
\method{print}{select_motifs}(x, ...)
}
\arguments{
\item{x}{An object of class "select_motifs"}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Method for select_motifs Objects
}
