% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models_rf_xgboost.R
\name{print.train_models_rf_xgboost}
\alias{print.train_models_rf_xgboost}
\title{Print method for train_models_rf_xgboost objects}
\usage{
\method{print}{train_models_rf_xgboost}(x, ...)
}
\arguments{
\item{x}{Object of class train_models_rf_xgboost.}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for train_models_rf_xgboost objects
}
