% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{sort.kmer_counts}
\alias{sort.kmer_counts}
\title{Sort method for kmer_counts objects}
\usage{
\method{sort}{kmer_counts}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{A kmer_counts object}

\item{decreasing}{Logical. Sort in decreasing order?}

\item{...}{Additional arguments (ignored)}
}
\description{
Sort method for kmer_counts objects
}
