% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{[.kmer_counts}
\alias{[.kmer_counts}
\title{Subset method for kmer_counts objects}
\usage{
\method{[}{kmer_counts}(x, i, ...)
}
\arguments{
\item{x}{A kmer_counts object}

\item{i}{Index or name for subsetting}

\item{...}{Additional arguments (ignored)}
}
\description{
Subset method for kmer_counts objects
}
