% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cluster_motifs.R
\name{summary.cluster_motifs}
\alias{summary.cluster_motifs}
\title{Summary Method for Cluster Motifs}
\usage{
\method{summary}{cluster_motifs}(object, ...)
}
\arguments{
\item{object}{Object of class 'cluster_motifs'}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for Cluster Motifs
}
