% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{summary.kmer_counts}
\alias{summary.kmer_counts}
\title{Summary method for kmer_counts objects}
\usage{
\method{summary}{kmer_counts}(object, ...)
}
\arguments{
\item{object}{A kmer_counts object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for kmer_counts objects
}
