% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_motifs.R
\name{summary.select_motifs}
\alias{summary.select_motifs}
\title{Summary Method for select_motifs Objects}
\usage{
\method{summary}{select_motifs}(object, ...)
}
\arguments{
\item{object}{An object of class "select_motifs"}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns object
}
\description{
Summary Method for select_motifs Objects
}
