% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models_rf_xgboost.R
\name{summary.train_models_rf_xgboost}
\alias{summary.train_models_rf_xgboost}
\title{Summary method for train_models_rf_xgboost objects}
\usage{
\method{summary}{train_models_rf_xgboost}(object, ...)
}
\arguments{
\item{object}{Object of class train_models_rf_xgboost.}

\item{...}{Additional arguments (unused).}
}
\description{
Summary method for train_models_rf_xgboost objects
}
