% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_kmers.R
\name{tail.kmer_counts}
\alias{tail.kmer_counts}
\title{Tail method for kmer_counts objects}
\usage{
\method{tail}{kmer_counts}(x, n = 6L, ...)
}
\arguments{
\item{x}{A kmer_counts object}

\item{n}{Number of elements to return}

\item{...}{Additional arguments (ignored)}
}
\description{
Tail method for kmer_counts objects
}
